requirements.txt
aiogram>=3.14.0
aiohttp>=3.10.0
python-dotenv>=1.0.0
cryptography>=46.0.0

# Polydictor - Twitter Intelligence
tweepy>=4.14.0
websocket-client>=1.6.0

# Solana blockchain integration
solana>=0.30.0
solders>=0.18.0
base58>=2.1.1
spl-token>=0.2.0
